+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 ..
     This file is a part of Kurt McKee's GitHub Workflows project.
     https://github.com/kurtmckee/github-workflows
-    Copyright 2024 Kurt McKee <contactme@kurtmckee.org>
+    Copyright 2024-2025 Kurt McKee <contactme@kurtmckee.org>
     SPDX-License-Identifier: MIT
 
 
@@ -16,6 +16,23 @@ Unreleased changes to the code are documented in
 in the ``changelog.d/`` directory on GitHub.
 
 ..  scriv-insert-here
+
+.. _changelog-1.1:
+
+1.1 — 2025-01-27
+================
+
+Added
+-----
+
+-   Add a new reusable workflow: ``create-pr.yaml``.
+
+    It is capable of running a tox label, committing all changes, and opening a PR.
+
+Fixed
+-----
+
+-   ``tox``: Remove a duplicate write to ``.tox-config.json``.
 
 .. _changelog-1.0: